refactor utils to be modules
