+++ modules/CredentialManager.py
